Re-attach the source volume before copying.
For some reason that is not entirely clear to me, after running some programs (for example, xar),
the creation dates *in the source* are modified. To prevent this or other potential modifications
of the source volumes when copying or cloning, it is necessary to re-attach
the source volume in read-only mode. Only remounting the volume in read-only mode
does not seem sufficient (why?).